make header only

+++ main.cpp
@@ -1,4 +1,4 @@
-#include "src/pack.h"
+#include "include/pack.h"
 #include <cstring>
 #include <iostream>
 
@@ -6,7 +6,7 @@ int main()
 {
     std::cout << "Trying PHP pack() in c++\n";
 
-    int x = 65;
+    long long x = 65;
     auto result = PhPacker::pack('q', x);
 
     std::cout << "RESULT: " << result << std::endl;
